+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/properties/events/MPROnHit.java
@@ -38,10 +38,7 @@ public class MPROnHit extends MPREvent {
 		if (this.target == null)
 			throw new JsonValidationException("Missing \"target\" for OnHit object: %s".formatted(this));
 
-		if (this.potionEffects == null) {
-			throw new JsonValidationException("Missing \"potion_effects\" for OnHit object: %s".formatted(this));
-		}
-		else {
+		if (this.potionEffects != null) {
 			for (MPRPotionEffect potionEffect : this.potionEffects)
 				potionEffect.validate();
 		}
@@ -78,15 +75,19 @@ public class MPROnHit extends MPREvent {
 		}
 
 		if (this.target == Target.ENTITY) {
-			for (MPRPotionEffect potionEffect : this.potionEffects) {
-				potionEffect.apply(entity, entity.level);
+			if (this.potionEffects != null) {
+				for (MPRPotionEffect potionEffect : this.potionEffects) {
+					potionEffect.apply(entity, entity.level);
+				}
 			}
 			this.tryPlaySound(entity);
 			this.tryExecuteFunction(entity);
 		}
 		else if (this.target == Target.OTHER) {
-			for (MPRPotionEffect potionEffect : this.potionEffects) {
-				potionEffect.apply(other, other.level);
+			if (this.potionEffects != null) {
+				for (MPRPotionEffect potionEffect : this.potionEffects) {
+					potionEffect.apply(other, other.level);
+				}
 			}
 			this.tryPlaySound(other);
 			this.tryExecuteFunction(other);